requirements.txt
Django~=3.1.7
djangorestframework~=3.12.2
djangorestframework-simplejwt~=4.6.0